flask-api/requirements.txt
blinker
click
cmake
filelock
Flask
Flask-Cors
itsdangerous
Jinja2
joblib
lit
MarkupSafe
mpmath
networkx
numpy
nvidia-cublas-cu11
nvidia-cuda-cupti-cu11
nvidia-cuda-nvrtc-cu11
nvidia-cuda-runtime-cu11
nvidia-cudnn-cu11
nvidia-cufft-cu11
nvidia-curand-cu11
nvidia-cusolver-cu11
nvidia-cusparse-cu11
nvidia-nccl-cu11
nvidia-nvtx-cu11
opencv-python
Pillow
sympy
torch
triton
typing_extensions
Werkzeug
